Extract code shared between sources to common/source
